requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
dbfread==2.0.7
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
future==0.18.3
GDAL @ file:///D:/Projects%20BGIS/Dependecies/GDAL-3.4.3-cp310-cp310-win_amd64.whl
geoserver-restconfig==2.0.8
gisdata==0.5.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
pygdaltools==1.3.2
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
sridentify==0.8.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
